code/requirements.txt
Charm-Crypto==0.50
cryptography==3.4.8
pycryptodome==3.22.0